daemon: Fail agent startup on incompatible datapath mode
Cilium does not currently support migrating a live node with pods between `netkit` and `veth` datapath modes
([ref](https://docs.cilium.io/en/stable/operations/performance/tuning/#netkit-device-mode)). But there is currently no safety mechanisms to safeguard against it. This means that is it currently possible to accidentally switch datapath mode on a node and end up in an undefined and unexpected state.

This PR aims to ensure this can not happen by adding a check in the endpoint restore logic to check the restored endpoint link type against the configured datapath mode and make the agent crash in case of incompatibility.

Signed-off-by: Hadrien Patte <hadrien.patte@datadoghq.com>